+++ e3nn/nn/__init__.py
@@ -1,3 +1,4 @@
+from .cut import Cut  # noqa
 from .batchnorm import *  # noqa
 from .fc import *  # noqa
 from .gate import *  # noqa