+++ ChangeLog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Module `Cut` to cut irreps tensors in sub tensors
 - Recursive TorchScript compiler `e3nn.util.jit`
 - TorchScript support for `TensorProduct` and subclasses, `NormActivation`, `Gate`, `FullyConnectedNet`, and `gate_points_2101.Network`
 ### Removed